Add a workflow that applies type information based on Objective-C runtime calls
The initial focus is on adjusting the call types of the objc_alloc-like methods to have a return type that is the struct matching the class object that is passed to the call.

This does not work well with the shared cache at present due to the interaction between inlining and call type adjustments (https://github.com/Vector35/binaryninja-api/issues/6737).