WIP: Apply return types to objc_msgSend when selector is in the init family
If the receiver is of a known type (not `id` or `objc_object*`), its return
type is applied to the `objc_msgSend` call.

This does not appear to be beneficial at present. I can see the call
type being overridden, but the inferred variable type does not change.
Perhaps this is https://github.com/Vector35/binaryninja-api/issues/6812?